Card-838 Error breaks firefox. Move bundle services out of Rule service, defer request to obtain bundles until publish dialog is opened. Also moves User request to a new User endpoint that respects auth header.
