+++ src/api/rule-engine/Rule.ts
@@ -9,6 +9,7 @@ import {
 } from "../system/http-response-util";
 import {ServerSideFieldModel, ServerSideTypeModel} from "./ServerSideFieldModel";
 import {I18nService} from "../system/locale/I18n";
+import {CoreWebService} from "../services/core-web-service";
 
 
 export const RULE_CREATE = 'RULE_CREATE'
@@ -56,23 +57,6 @@ export interface IRecord {
   deleting?:boolean
   errors?:any
   set?(string, any):any
-}
-
-export interface IUser {
-  firstName?: string,
-  lastName?: string,
-  roleId?: string,
-  userId?: string
-}
-
-export interface IBundle {
-  name?: string,
-  id?: string
-}
-
-export interface IPublishEnvironment {
-  name?: string,
-  id?: string
 }
 
 export interface IRuleAction extends IRecord {
@@ -255,16 +239,11 @@ export const DEFAULT_RULE:IRule = {
 }
 
 @Injectable()
-export class RuleService {
+export class RuleService extends CoreWebService {
   private _rulesEndpointUrl:string
   private _actionsEndpointUrl:string
   private _conditionTypesEndpointUrl:string
   private _ruleActionTypesEndpointUrl:string
-  private _bundleStoreUrl:string
-  private _loggedUserUrl:string
-  private _addToBundleUrl:string
-  private _pushEnvironementsUrl:string
-  private _pushRuleUrl:string
 
 
   ruleActionTypes$:BehaviorSubject<ServerSideTypeModel[]> = new BehaviorSubject([]);
@@ -279,21 +258,12 @@ export class RuleService {
   _conditionTypes:{[key:string]:ServerSideTypeModel} = {}
   private _conditionTypesAry:ServerSideTypeModel[] = []
 
-  bundles$:BehaviorSubject<IBundle[]> = new BehaviorSubject([]);
-  environments$:BehaviorSubject<IDBEnvironment[]> = new BehaviorSubject([]);
-  private _bundlesAry:IBundle[] = []
-  private _environmentsAry:IDBEnvironment[] = []
-
-  constructor(private _apiRoot:ApiRoot, private _http:Http, private _resources:I18nService) {
+  constructor(public _apiRoot:ApiRoot, _http:Http, private _resources:I18nService) {
+    super(_apiRoot, _http)
     this._rulesEndpointUrl = `${this._apiRoot.defaultSiteUrl}/ruleengine/rules`
     this._actionsEndpointUrl = `${this._apiRoot.defaultSiteUrl}/ruleengine/actions`
     this._conditionTypesEndpointUrl = `${this._apiRoot.baseUrl}api/v1/system/ruleengine/conditionlets`
     this._ruleActionTypesEndpointUrl = `${this._apiRoot.baseUrl}api/v1/system/ruleengine/actionlets`
-    this._bundleStoreUrl = `${this._apiRoot.baseUrl}api/bundle/getunsendbundles/userid` //TODO: dotcms.org.1 fmontes get this user id dinamically
-    this._loggedUserUrl = `${this._apiRoot.baseUrl}api/user/getloggedinuser/`
-    this._addToBundleUrl = `${this._apiRoot.baseUrl}DotAjaxDirector/com.dotcms.publisher.ajax.RemotePublishAjaxAction/cmd/addToBundle`
-    this._pushEnvironementsUrl = `${this._apiRoot.baseUrl}api/environment/loadenvironments/roleId`
-    this._pushRuleUrl = `${this._apiRoot.baseUrl}DotAjaxDirector/com.dotcms.publisher.ajax.RemotePublishAjaxAction/cmd/publish`
 
 
     this._preCacheCommonResources(_resources)
@@ -402,101 +372,6 @@ export class RuleService {
     }).map(RuleService.fromServerServersideTypesTransformFn)
   }
 
-  addRuleToBundle(ruleId:string, bundle:IBundle):Observable<{errorMessages:string[],total:number,errors:number}> {
-    return this.request({
-      body: `assetIdentifier=${ruleId}&bundleName=${bundle.name}&bundleSelect=${bundle.id}`,
-      method: RequestMethod.Post,
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      url: this._addToBundleUrl
-    })
-  }
-
-  getLoggedUser():Observable<IUser> {
-    return this.request({
-      method: RequestMethod.Get,
-      url: this._loggedUserUrl,
-    })
-  }
-
-  loadBundleStores(){
-    let obs
-    if (this._bundlesAry.length) {
-      obs = Observable.fromArray(this._bundlesAry)
-    } else {
-      obs = this._doLoadBundleStores().map((bundles:IBundle[])=> {
-        this._bundlesAry = bundles
-        console.log("RuleService", "loadBundles", bundles)
-        return bundles
-      })
-    }
-    obs.subscribe((bundles) => this.bundles$.next(bundles))
-  }
-
-  _doLoadBundleStores():Observable<IBundle[]> {
-    return this.getLoggedUser().flatMap((user:IUser) => {
-      return this.request({
-        method: RequestMethod.Get,
-        url: `${this._bundleStoreUrl}/${user.userId}`,
-      }).map(RuleService.fromServerBundleTransformFn)
-    })
-  }
-
-  loadPublishEnvironments() {
-    let obs
-    if (this._environmentsAry.length) {
-      obs = Observable.fromArray(this._environmentsAry)
-    } else {
-      obs = this._doLoadPublishEnvironments().map((environments:IDBEnvironment[])=> {
-        this._environmentsAry = environments
-        return environments
-      })
-    }
-    obs.subscribe((environments) => this.environments$.next(environments))
-  }
-
-  _doLoadPublishEnvironments():Observable<IPublishEnvironment[]> {
-    return this.getLoggedUser().flatMap((user:IUser) => {
-      return this.request({
-        method: RequestMethod.Get,
-        url: `${this._pushEnvironementsUrl}/${user.roleId}/?name=0`,
-      }).map(RuleService.fromServerEnvironmentTransformFn)
-    })
-  }
-
-  private getFormattedDate(date:Date) {
-    let month = (date.getMonth() + 1).toString()
-    month += month.length < 2 ? "0" + month : month
-    return `${month}-${date.getDate()}-${date.getFullYear()}`
-  }
-
-  private getPublishRuleData(ruleId:string, environmentId:string) {
-    let resul:string = "";
-    resul += `assetIdentifier=${ruleId}`
-    resul += `&remotePublishDate=${this.getFormattedDate(new Date())}`
-    resul += "&remotePublishTime=00-00"
-    resul += `&remotePublishExpireDate=${this.getFormattedDate(new Date())}`
-    resul += "&remotePublishExpireTime=00-00"
-    resul += "&iWantTo=publish"
-    resul += `&whoToSend=${environmentId}`
-    resul += "&bundleName="
-    resul += "&bundleSelect="
-    resul += "&forcePush=false"
-    return resul
-  }
-
-  pushPublishRule(ruleId:string, environmentId:string):Observable<{errorMessages:string[],total:number,bundleId:string,errors:number}> {
-    return this.request({
-      body: this.getPublishRuleData(ruleId, environmentId),
-      method: RequestMethod.Post,
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      url: this._pushRuleUrl
-    })
-  }
-
   private loadConditionTypes():Observable<ServerSideTypeModel[]> {
     let obs
     if (this._conditionTypesAry.length) {
@@ -520,46 +395,6 @@ export class RuleService {
       method: RequestMethod.Get,
       url: this._conditionTypesEndpointUrl,
     }).map(RuleService.fromServerServersideTypesTransformFn)
-  }
-
-  request(options:any):Observable<any> {
-    let headers:Headers = this._apiRoot.getDefaultRequestHeaders()
-    let tempHeaders = options.headers ? options.headers : {"Content-Type": "application/json"}
-    Object.keys(tempHeaders).forEach((key)=>{
-      headers.set(key, tempHeaders[key])
-    })
-    var source = options.body
-    if (options.body) {
-      if (typeof options.body !== 'string') {
-        options.body = JSON.stringify(options.body);
-      }
-    }
-    options.headers = headers
-
-    var request = new Request(options)
-    return this._http.request(request)
-        .map((resp:Response) => {
-          return hasContent(resp) ? resp.json() : resp
-        })
-        .catch((response:Response, original:Observable<any>):Observable<any> => {
-          if (response) {
-            if (response.status === 500) {
-              if (response.text() && response.text().indexOf("ECONNREFUSED") >= 0) {
-                throw new CwError(NETWORK_CONNECTION_ERROR, CLIENTS_ONLY_MESSAGES[NETWORK_CONNECTION_ERROR], request, response, source)
-              } else {
-                throw new CwError(SERVER_RESPONSE_ERROR, response.headers.get('error-message'), request, response, source)
-              }
-            }
-            else if (response.status === 404) {
-              console.error("Could not execute request: 404 path not valid.", options.url)
-              throw new CwError(UNKNOWN_RESPONSE_ERROR, response.headers.get('error-message'), request, response, source)
-            } else {
-              console.log("Could not execute request: Response status code: ", response.status, 'error:', response, options.url)
-              throw new CwError(UNKNOWN_RESPONSE_ERROR, response.headers.get('error-message'), request, response, source)
-            }
-          }
-          return null
-        })
   }
 
   static fromServerRulesTransformFn(ruleMap):RuleModel[] {
@@ -615,15 +450,6 @@ export class RuleService {
     }
   }
 
-  static fromServerBundleTransformFn(data):IBundle[] {
-    return data.items || [];
-  }
-
-  static fromServerEnvironmentTransformFn(data):IPublishEnvironment[] {
-    // Endpoint return extra empty environment
-    data.shift()
-    return data
-  }
 
   static fromServerServersideTypesTransformFn(typesMap):ServerSideTypeModel[] {
     let types = Object.keys(typesMap).map((key:string) => {
