+++ src/view/components/common/push-publish/add-to-bundle-dialog-container.ts
@@ -1,15 +1,15 @@
 import {Component, ChangeDetectionStrategy, Input, Output, EventEmitter} from "angular2/core";
 import {CORE_DIRECTIVES} from "angular2/common";
-import {IBundle, RuleService} from "../../../../api/rule-engine/Rule";
 import {AddToBundleDialogComponent} from "./add-to-bundle-dialog-component";
 import {BehaviorSubject} from "rxjs/Rx";
+import {BundleService, IBundle} from "../../../../api/services/bundle-service";
 
 @Component({
   selector: 'cw-add-to-bundle-dialog-container',
   directives: [CORE_DIRECTIVES, AddToBundleDialogComponent],
   template: `
   <cw-add-to-bundle-dialog-component
-  [bundleStores]="ruleService.bundles$ | async"
+  [bundleStores]="bundleService.bundles$ | async"
   [hidden]="hidden"
   [errorMessage]="errorMessage | async"
   (cancel)="hidden = true; close.emit($event); errorMessage = null;"
@@ -25,21 +25,22 @@ export class AddToBundleDialogContainer {
   @Output() cancel:EventEmitter<boolean> = new EventEmitter(false)
 
   errorMessage:BehaviorSubject<string> = new BehaviorSubject(null)
-
-  bundleStores:IBundle[] = []
-
-  constructor(public ruleService:RuleService) {
-    this.ruleService.loadBundleStores()
+  bundlesLoaded:boolean = false
+  
+  constructor(public bundleService:BundleService) {
+    
   }
 
   ngOnChanges(change){
-    if (change.hidden) {
+    if (change.hidden && !this.hidden && !this.bundlesLoaded) {
       console.log("AddToBundlehDialogContainer", "ngOnChanges", change.hidden.currentValue, change.hidden.previousValue)
+      this.bundlesLoaded = true
+      this.bundleService.loadBundleStores()
     }
   }
 
   addToBundle(bundle:IBundle) {
-    this.ruleService.addRuleToBundle(this.assetId, bundle).subscribe((result:any)=> {
+    this.bundleService.addRuleToBundle(this.assetId, bundle).subscribe((result:any)=> {
       if (!result.errors) {
         this.close.emit({isCanceled:false})
         this.errorMessage = null
