+++ tsconfig.json
@@ -26,6 +26,9 @@
     "./src/api/rule-engine/Rule.spec.ts",
     "./src/api/rule-engine/Rule.ts",
     "./src/api/rule-engine/ServerSideFieldModel.ts",
+    "./src/api/services/bundle-service.it-spec.ts",
+    "./src/api/services/bundle-service.ts",
+    "./src/api/services/core-web-service.ts",
     "./src/api/system/http-response-util.ts",
     "./src/api/system/locale/I18n.it-spec.ts",
     "./src/api/system/locale/I18n.ts",
@@ -46,10 +49,11 @@
     "./src/e2e/view/rule-engine/rule-engine-page.ts",
     "./src/view/boot.ts",
     "./src/view/components/common/google-map/area-picker-dialog.component.ts",
-    "./src/view/components/common/google-map/area-picker-dialog.container.ts",
     "./src/view/components/common/google-map/boot.ts",
     "./src/view/components/common/google-map/demo.ts",
     "./src/view/components/common/modal-dialog/dialog-component.ts",
+    "./src/view/components/common/push-publish/add-to-bundle-dialog-component.ts",
+    "./src/view/components/common/push-publish/add-to-bundle-dialog-container.ts",
     "./src/view/components/common/push-publish/push-publish-dialog-component.ts",
     "./src/view/components/common/push-publish/push-publish-dialog-container.ts",
     "./src/view/components/input/toggle/demo.ts",
